+++ packages/lumx-react/src/components/date-picker/DatePicker.tsx
@@ -56,6 +56,7 @@ const DEFAULT_PROPS: Partial<DatePickerProps> = {
 
 /**
  * Simple component used to pick a date (semi-controlled implementation).
+ * @param props See DatePickerProps.
  *
  * @return The component.
  */
@@ -66,12 +67,15 @@ const DatePicker = (props: DatePickerProps) => {
     const setPrevMonth = () => setMonthOffset(monthOffset - 1);
     const setNextMonth = () => setMonthOffset(monthOffset + 1);
 
+    const selectedMonth = moment(today)
+        .locale(props.locale)
+        .add(monthOffset, 'months');
+
     return (
         <DatePickerControlled
-            monthOffset={monthOffset}
-            today={today}
             onPrevMonthChange={setPrevMonth}
             onNextMonthChange={setNextMonth}
+            selectedMonth={selectedMonth}
             {...props}
         />
     );