+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   Added optional `defaultMonth` prop to components `DatePickerField` and `DatePicker`.
+
+### Changed
+
+-   _[BREAKING]_ Replaced `today` and `monthOffset` props by `selectedMonth` for `DatePickerControlled` component.
+
 ## [0.22.5][] - 2020-06-05
 
 ### Fixed